spaces/nexora-image-gen/requirements.txt
gradio==4.7.1
Pillow==10.1.0
transformers==4.35.0
torch==2.1.0
diffusers==0.24.0
accelerate==0.25.0

